+++ esp-backtrace/Cargo.toml
@@ -43,6 +43,7 @@ print-float-registers = [] # TODO support esp32p4
 # You may optionally enable one or more of the below features to provide
 # additional functionality:
 colors               = []
+coredump             = []
 custom-halt          = []
 custom-pre-backtrace = []
 exception-handler    = []
